Move test result class to seperate file
