+++ app/src/main/java/com/example/lanproject/TaskFragment1.kt
@@ -17,15 +17,6 @@ import org.json.JSONObject
 import java.text.SimpleDateFormat
 import java.time.LocalDateTime
 
-data class TestResult(
-    var result : Int,
-    var maxResult : Int,
-    val difficulty : Int, // TODO(lucas): Of which type should we represent the difficulty?
-    val timeStamp : Long,
-) {
-
-}
-
 class TaskFragment1 : Fragment(R.layout.fragment_task1) {
 
     private lateinit var inflater: LayoutInflater
